Made iface_idx an int, since that\'s what is being used everywhere

+++ hw3/odr.h
@@ -30,11 +30,11 @@ typedef struct eth_frame {
 
 // TODO This needs to be filled correctly
 typedef struct route_entry {
-  char ip_addr[16];            // The IP address of the machine we are maintaining this entry for. i.e. The IP address of the final destination
-  uint16_t iface_idx;          // The interface index through which we reach the next hop
-  char next_hop[6];            // The ethernet address of the next hop
-  uint16_t nhops_to_dest;      // Number of hops to destination
-  uint32_t last_updated_at_ms; // When was this entry updated?
+  char ip_addr[16];             // The IP address of the machine we are maintaining this entry for. i.e. The IP address of the final destination
+  int iface_idx;                // The interface index through which we reach the next hop
+  char next_hop[6];             // The ethernet address of the next hop
+  uint16_t nhops_to_dest;       // Number of hops to destination
+  uint32_t last_updated_at_ms;  // When was this entry updated?
 } route_entry;
 
 // This is the entry used to forward messages from the
